requirements.txt
absl-py==0.15.0
alignfaces @ git+https://git@github.com/SourCherries/auto-face-align.git@365bd01c22da6f3a44190261786fcc585687ea50
antlr4-python3-runtime==4.8
appnope==0.1.2
asttokens==2.0.5
astunparse==1.6.3
backcall==0.2.0
better-exceptions==0.3.3
black==22.1.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
click==8.0.3
cmake==3.21.3
cycler==0.10.0
decorator==5.1.1
dlib==19.22.1
executing==0.8.2
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.1
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.1
h5py==3.1.0
idna==3.3
imageio==2.9.0
imutils==0.5.4
ipdb==0.13.9
ipython==8.0.1
jedi==0.18.1
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lime==0.2.0.1
Markdown==3.3.4
matplotlib==3.4.3
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
networkx==2.6.3
numpy==1.19.5
oauthlib==3.1.1
omegaconf==2.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.4
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.0
prompt-toolkit==3.0.28
protobuf==3.19.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.11.2
pyparsing==3.0.1
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.3
scikit-learn==1.0.1
scipy==1.7.1
seaborn==0.11.2
Shapely==1.8.1.post1
simplejson==3.17.6
six==1.15.0
stack-data==0.1.4
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
threadpoolctl==3.0.0
tifffile==2021.10.12
toml==0.10.2
tomli==2.0.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
Werkzeug==2.0.2
wrapt==1.12.1
